Added Relation and UserAccount repositories along with a few changes in the UserAccount and Relation classes.
In the UserAccount class, a new constructor was added for easier object creation. The change in the Relation model was regarding the type of 'createdAt' field from Timestamp to LocalDateTime for a more efficient datetime handling in Java. Also added a new method in Relation constructor to initialize 'relationId' with respective user id's.

Created new RelationRepository and UserAccountRepository classes to facilitate the CRUD operations on Relation and UserAccount objects. Also added initial data in the main application class (PayMyBuddyAppApplication) for testing purposes.